refactor: #8 履歴の一覧表示で使用していたものを親コンポーネントに記述
